requirements.txt
numpy
pandas
scikit-learn
python-dotenv
pymysql
mysql-connector-python
dvc
SQLAlchemy